# internal/serverapi/openapi_gen.go
// Package serverapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package serverapi

import (
	"errors"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// CreateEnrollmentTokenRequest defines model for CreateEnrollmentTokenRequest.
type CreateEnrollmentTokenRequest struct {
	Description *string `json:"description,omitempty"`
}

// EnrollmentToken defines model for EnrollmentToken.
type EnrollmentToken struct {
	CreatedAt   *time.Time `json:"created_at,omitempty"`
	Description *string    `json:"description,omitempty"`
	ExpiresAt   time.Time  `json:"expires_at"`
	Id          string     `json:"id"`

	// Token Plaintext token; present only on creation
	Token  *string    `json:"token,omitempty"`
	UsedAt *time.Time `json:"used_at,omitempty"`
}

// Error defines model for Error.
type Error struct {
	// Code Stable machine-readable error code, e.g. node.occupied, plan.incompatible
	Code    string                  `json:"code"`
	Details *map[string]interface{} `json:"details,omitempty"`
	Message string                  `json:"message"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// MigrationImportRequest defines model for MigrationImportRequest.
type MigrationImportRequest struct {
	Confirm        interface{} `json:"confirm"`
	IniPath        *string     `json:"ini_path,omitempty"`
	LegacyRevision *string     `json:"legacy_revision,omitempty"`
	LegacyRoot     string      `json:"legacy_root"`
	LegacyState    string      `json:"legacy_state"`
	PlanDigest     string      `json:"plan_digest"`
}

// MigrationScanRequest defines model for MigrationScanRequest.
type MigrationScanRequest struct {
	Docker         *bool   `json:"docker,omitempty"`
	IniPath        *string `json:"ini_path,omitempty"`
	LegacyRevision *string `json:"legacy_revision,omitempty"`
	LegacyRoot     string  `json:"legacy_root"`
	LegacyState    string  `json:"legacy_state"`
}

// Module defines model for Module.
type Module struct {
	ArtifactKinds *[]string `json:"artifactKinds,omitempty"`
	Capabilities  *[]string `json:"capabilities,omitempty"`
	Id            string    `json:"id"`
	JobKinds      *[]string `json:"jobKinds,omitempty"`
	Nav           *struct {
		Icon  *string `json:"icon,omitempty"`
		Label *string `json:"label,omitempty"`
		Order *int    `json:"order,omitempty"`
	} `json:"nav,omitempty"`
	Route          string                  `json:"route"`
	SettingsSchema *map[string]interface{} `json:"settingsSchema,omitempty"`
	Title          string                  `json:"title"`
}

// RunAccepted defines model for RunAccepted.
type RunAccepted struct {
	RunId openapi_types.UUID `json:"run_id"`
}

// Session defines model for Session.
type Session struct {
	CsrfToken string    `json:"csrf_token"`
	ExpiresAt time.Time `json:"expires_at"`
	Username  string    `json:"username"`
}

// SystemInfo defines model for SystemInfo.
type SystemInfo struct {
	AgentUrl      string `json:"agent_url"`
	CaFingerprint string `json:"ca_fingerprint"`
	Commit        string `json:"commit"`
	Nodes         struct {
		Approved *int `json:"approved,omitempty"`
		Online   *int `json:"online,omitempty"`
		Total    *int `json:"total,omitempty"`
	} `json:"nodes"`
	Time    time.Time `json:"time"`
	Version string    `json:"version"`
}

// ID defines model for ID.
type ID = string

// Conflict defines model for Conflict.
type Conflict = Error

// NotFound defines model for NotFound.
type NotFound = Error

// ServiceUnavailable defines model for ServiceUnavailable.
type ServiceUnavailable = Error

// Unauthorized defines model for Unauthorized.
type Unauthorized = Error

// Unprocessable defines model for Unprocessable.
type Unprocessable = Error

// StreamEventsParams defines parameters for StreamEvents.
type StreamEventsParams struct {
	// Types Comma-separated event type filter
	Types *string `form:"types,omitempty" json:"types,omitempty"`
}

// CreateEnrollmentTokenJSONRequestBody defines body for CreateEnrollmentToken for application/json ContentType.
type CreateEnrollmentTokenJSONRequestBody = CreateEnrollmentTokenRequest

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = LoginRequest

// MigrationImportJSONRequestBody defines body for MigrationImport for application/json ContentType.
type MigrationImportJSONRequestBody = MigrationImportRequest

// MigrationScanJSONRequestBody defines body for MigrationScan for application/json ContentType.
type MigrationScanJSONRequestBody = MigrationScanRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /enrollment-tokens)
	ListEnrollmentTokens(w http.ResponseWriter, r *http.Request)
	// CreateEnrollmentToken Create a one-use, ten-minute enrollment token
	// (POST /enrollment-tokens)
	CreateEnrollmentToken(w http.ResponseWriter, r *http.Request)

	// (DELETE /enrollment-tokens/{id})
	DeleteEnrollmentToken(w http.ResponseWriter, r *http.Request, id ID)
	// StreamEvents Server-sent event stream (telemetry, job state, log tails)
	// (GET /events)
	StreamEvents(w http.ResponseWriter, r *http.Request, params StreamEventsParams)
	// Login Exchange username/password for a session cookie
	// (POST /login)
	Login(w http.ResponseWriter, r *http.Request)
	// Logout End the current session
	// (POST /logout)
	Logout(w http.ResponseWriter, r *http.Request)

	// (GET /metrics)
	PrometheusMetrics(w http.ResponseWriter, r *http.Request)
	// MigrationImport Start a digest-gated import into an isolated staging state root
	// (POST /migration/import)
	MigrationImport(w http.ResponseWriter, r *http.Request)
	// MigrationScan Start a resumable legacy DGX-Dashboard scan job
	// (POST /migration/scan)
	MigrationScan(w http.ResponseWriter, r *http.Request)
	// ListModules Enabled first-party modules (descriptors present server- and client-side)
	// (GET /modules)
	ListModules(w http.ResponseWriter, r *http.Request)
	// GetSession Current session metadata including the CSRF token
	// (GET /session)
	GetSession(w http.ResponseWriter, r *http.Request)

	// (GET /system/info)
	SystemInfo(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// (GET /enrollment-tokens)
func (_ Unimplemented) ListEnrollmentTokens(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// CreateEnrollmentToken Create a one-use, ten-minute enrollment token
// (POST /enrollment-tokens)
func (_ Unimplemented) CreateEnrollmentToken(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (DELETE /enrollment-tokens/{id})
func (_ Unimplemented) DeleteEnrollmentToken(w http.ResponseWriter, r *http.Request, id ID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// StreamEvents Server-sent event stream (telemetry, job state, log tails)
// (GET /events)
func (_ Unimplemented) StreamEvents(w http.ResponseWriter, r *http.Request, params StreamEventsParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Login Exchange username/password for a session cookie
// (POST /login)
func (_ Unimplemented) Login(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Logout End the current session
// (POST /logout)
func (_ Unimplemented) Logout(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /metrics)
func (_ Unimplemented) PrometheusMetrics(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// MigrationImport Start a digest-gated import into an isolated staging state root
// (POST /migration/import)
func (_ Unimplemented) MigrationImport(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// MigrationScan Start a resumable legacy DGX-Dashboard scan job
// (POST /migration/scan)
func (_ Unimplemented) MigrationScan(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ListModules Enabled first-party modules (descriptors present server- and client-side)
// (GET /modules)
func (_ Unimplemented) ListModules(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// GetSession Current session metadata including the CSRF token
// (GET /session)
func (_ Unimplemented) GetSession(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /system/info)
func (_ Unimplemented) SystemInfo(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ListEnrollmentTokens operation middleware
func (siw *ServerInterfaceWrapper) ListEnrollmentTokens(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListEnrollmentTokens(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateEnrollmentToken operation middleware
func (siw *ServerInterfaceWrapper) CreateEnrollmentToken(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateEnrollmentToken(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteEnrollmentToken operation middleware
func (siw *ServerInterfaceWrapper) DeleteEnrollmentToken(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id ID

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: r.URL.RawPath == ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteEnrollmentToken(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// StreamEvents operation middleware
func (siw *ServerInterfaceWrapper) StreamEvents(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params StreamEventsParams

	// ------------- Optional query parameter "types" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "types", r.URL.Query(), &params.Types, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "types"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "types", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.StreamEvents(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Login operation middleware
func (siw *ServerInterfaceWrapper) Login(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Login(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Logout operation middleware
func (siw *ServerInterfaceWrapper) Logout(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Logout(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PrometheusMetrics operation middleware
func (siw *ServerInterfaceWrapper) PrometheusMetrics(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PrometheusMetrics(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// MigrationImport operation middleware
func (siw *ServerInterfaceWrapper) MigrationImport(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.MigrationImport(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// MigrationScan operation middleware
func (siw *ServerInterfaceWrapper) MigrationScan(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.MigrationScan(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListModules operation middleware
func (siw *ServerInterfaceWrapper) ListModules(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListModules(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetSession operation middleware
func (siw *ServerInterfaceWrapper) GetSession(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetSession(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// SystemInfo operation middleware
func (siw *ServerInterfaceWrapper) SystemInfo(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SystemInfo(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/login", wrapper.Login)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/logout", wrapper.Logout)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/session", wrapper.GetSession)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/system/info", wrapper.SystemInfo)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/metrics", wrapper.PrometheusMetrics)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/modules", wrapper.ListModules)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/events", wrapper.StreamEvents)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/enrollment-tokens", wrapper.ListEnrollmentTokens)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/enrollment-tokens", wrapper.CreateEnrollmentToken)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/enrollment-tokens/{id}", wrapper.DeleteEnrollmentToken)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/migration/scan", wrapper.MigrationScan)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/migration/import", wrapper.MigrationImport)
	})

	return r
}